seproject/seproject/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
psycopg2-binary>=2.9
django-cors-headers>=4.0
gunicorn>=20.1
